Message de notification en cas de commandes manquantes

+++ pdf.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-file=$(mktemp)
-unoconv -o $file "$1"
-mimeopen $file
-rm -f $file
+if ! which unoconv > /dev/null
+then
+  notify-send "$0: unoconv was not found"
+  exit 1
+else
+  file=$(mktemp)
+  unoconv -o $file "$1"
+  mimeopen -n $file
+  rm -f $file
+fi
